added functionality to the canvas for formation and play update and create and fetch logic

+++ Server/xno-backend/src/main/java/com/xno/xno_backend/services/PlayServiceImpl.java
@@ -202,8 +202,9 @@ public class PlayServiceImpl implements PlayService {
             String playImageUrl = uploadResult.get("secure_url").toString();
             String playPublicId = uploadResult.get("public_id").toString();
 
-            imageService.deleteImage(existing.getPlayPublicId());
-
+            if(playRepository.countByPlayImageUrlAndIdNot(existing.getPlayImageUrl(), existing.getPlayId()) == 0) {
+                imageService.deleteImage(existing.getPlayPublicId());
+            }
             existing.setPlayImageUrl(playImageUrl);
             existing.setPlayPublicId(playPublicId);
         }
